requirements.txt
python-amazon-sp-api>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.25.0
